Refactor imagenet_experiment file
