+++ projects/imagenet/imagenet_tune.py
@@ -29,7 +29,7 @@ from ray.tune import Trainable, tune
 from ray.tune.resources import Resources
 
 from experiments import CONFIGS
-from nupic.research.frameworks.pytorch.imagenet_experiment import ImagenetExperiment
+from nupic.research.frameworks.pytorch.imagenet import ImagenetExperiment
 
 os.environ["HDF5_USE_FILE_LOCKING"] = "FALSE"
